reliability: check for `sort` executable in `cheshire3-init`
In response to issue #70

Added a check for the existence of the Unix sort utility
executable when configuring a new databse using `cheshire3-init`.
If the `sortPath` configured at server level does not exist, try to
locate it using `which sort` and add the value to the configuration
for the default IndexStore.